+++ src/scripts/lib/Xlsx/files/index.d.ts
@@ -1,0 +1,4 @@
+export const contentTypes: XMLDocument;
+export const relationships: XMLDocument;
+export const workSheet: XMLDocument;
+export function workBook(sheetName?: string, sheetId?: string): {data: XMLDocument,rels: XMLDocument};