+++ src/scripts/lib/Xlsx/files/relationship.d.ts
@@ -1,0 +1,3 @@
+export default function generateRelationships(): XMLDocument;
+export default function generateRelationships(type: string): XMLDocument;
+export default function generateRelationships(tyoe: string, id?: string): XMLDocument;