+++ src/scripts/lib/Xlsx/modules/write.d.ts
@@ -1,0 +1,4 @@
+import { xlsxConfig } from "./validator"
+
+/** Create Xlsx file based on its config */
+export function createXlsx(config: xlsxConfig): void;