+++ src/scripts/models/CSVFile.js
@@ -1,7 +1,7 @@
 import MyArray from "./MyArray";
 import { rgx } from "../lib/Constant";
 import { isNumeric, createFilteredObject } from "../lib/Util";
-import readXlsxFile from "../lib/Xlsx";
+import readXlsxFile, { writeXlsx } from "../lib/Xlsx";
 import { options } from "./Settings";
 import * as util from "../lib/CSVUtil";
 
@@ -26,6 +26,7 @@ import * as util from "../lib/CSVUtil";
 
 class CSVFile {
     constructor() {
+        /** @type {MyArray<string>} */
         this.types = new MyArray("gagal", "error");
         /** @type {fileName} */
         this.fileName = { fname: [], ftype: "", numb: 1, ext: ".csv" };
@@ -41,7 +42,7 @@ class CSVFile {
         if (!file) return this;
         this.constructFileName(file.name);
         const data = await ((f, { xlsxFileCheck }, { ext }) => {
-            return xlsxFileCheck.test(f.type) && /(?:xlsx)/g.test(ext)
+            return xlsxFileCheck.test(f.type) && /xlsx/.test(ext)
                 ? f.arrayBuffer()
                 : f.text();
         })(file, rgx, this.fileName);
@@ -147,6 +148,7 @@ class CSVFile {
                         val: ["'", '"', `$1`][i],
                     };
                 })(i);
+                if (v === "") data.length++;
                 if (!!v) data.push(v.replace(rgx, val));
             });
             return "";
@@ -154,7 +156,7 @@ class CSVFile {
         text.replace(util.rowValue(delimiter, "g"), replacer);
 
         if (/,\s*$/.test(text)) data.push("");
-        return data.nonEmptyValue;
+        return data;
     }
 
     /**
@@ -305,7 +307,16 @@ class CSVFile {
         //     return new Blob([theData], { type: "text/csv;charset=utf-8;" });
         // })(data);
 
-        return { fileUrl: URL.createObjectURL(csvData), fileName: `${name}.csv` };
+        return { fileUrl: URL.createObjectURL(csvData), fileName: `${name}` };
+    }
+
+    /**
+     * Create a Xlsx file from array data
+     * @param {String} name
+     * @param {MyArray | Array} data
+     */
+    static exportToXlsx(name, data) {
+        return writeXlsx(data, name);
     }
 }