+++ src/scripts/models/Reports.d.ts
@@ -1,20 +1,26 @@
 import MyArray from "./MyArray";
 
+type reportData = {
+    data: MyArray<Object>;
+    count: number;
+    add(data: any): reportData;
+}
+
 export default class Reports {
-    sukses: number;
-    gagal: MyArray<Object>;
-    error: MyArray<Object>;
+    sukses: reportData;
+    gagal: reportData;
+    error: reportData;
 
     /** Reset report */
     reset(): void;
 
     /** Create report data for error and failed */
-    createData(type: "gagal" | "error"): { count: number; data: MyArray<any> };
+    createData(type: "sukses" | "gagal" | "error"): reportData;
 
     /** Add success record */
-    success(): void;
+    success(i: Object): void;
 
     /** Add fail record */
-    fail(i: number, err?: number): void;
+    fail(i: Object, err?: number): void;
 }
 export const report: Reports;