+++ src/scripts/models/Reports.js
@@ -1,45 +1,59 @@
 import MyArray from "./MyArray";
 
+class reportData {
+    constructor() {
+        this.data = new MyArray();
+    }
+    get count() {
+        return this.data.length;
+    }
+    add(data) {
+        this.data.push(data);
+        return this;
+    }
+}
+
 class Reports {
     constructor() {
-        this.sukses = 0;
-        this.gagal = new MyArray();
-        this.error = new MyArray();
+        this.sukses = new reportData();
+        this.gagal = new reportData();
+        this.error = new reportData();
     }
 
     /**
      * Reset report
      */
     reset() {
-        this.sukses = 0;
-        this.gagal = new MyArray();
-        this.error = new MyArray();
+        for (let type in this) {
+            this[type] = new reportData();
+        }
+        return this;
     }
 
     /**
      * Create report data for error and failed
-     * @param {"gagal" | "error"} type type report
-     * @returns {{count:number, data:MyArray}} report count with it's detail
+     * @param {"sukses" | "gagal" | "error"} type type report
+     * @returns {reportData} report count with it's detail
      */
     createData(type) {
-        let data = this[type] || new MyArray();
-        return { count: data.length, data: data };
+        return this.hasOwnProperty(type) ? this[type] : new reportData();
     }
 
     /**
      * Add success record
+     * @param {Object} data queue index
      */
-    success() {
-        this.sukses++;
+    success(data) {
+        this.sukses.add(data);
     }
 
     /**
      * add fail record
-     * @param {Number} i queue index
+     * @param {Object} i queue index
      * @param {Number} err error type
      */
     fail(i, err = 1) {
-        err ? this.error.push(i) : this.gagal.push(i);
+        err ? this.error.add(i) : this.gagal.add(i);
     }
 }